requirements.txt
connexion == 2.7.0
swagger-ui-bundle == 0.0.6
python_dateutil == 2.6.0
setuptools >= 21.0.0
Flask-Cors==3.0.8
google-cloud-datastore==1.10.0
flask_testing==0.6.1
google-cloud-profiler
PyJWT==1.7.1